requirements.txt
python-dotenv>=1.1.1
pydantic>=2.12.0
fastapi>=0.118.3
uvicorn>=0.37.0
openai>=2.3.0
tiktoken>=0.12.0
anthropic>=0.69.0
google-genai>=1.43.0
